pyproject.toml
[project]
name = "memory-service"
version = "0.1.0"
description = "Production-style memory service for AI agents using FastAPI, Postgres, Redis, and Vector DB."
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { text = "MIT" }
authors = [
  { name = "Gabriel Caramori", email = "gabrielcaramori@hotmail.com" }
]

dependencies = [
  "fastapi==0.104.1",
  "uvicorn[standard]==0.24.0",

  # Database
  "sqlalchemy==2.0.23",
  "psycopg2-binary==2.9.9",

  # Cache
  "redis==5.0.1",

  # Vector DB
  "chromadb==1.3.5",

  # Embeddings
  "sentence-transformers==2.7.0",

  # Utils
  "numpy==1.26.4",
  "python-dotenv==1.0.1",
  "httpx==0.28.1",
  "pydantic-settings==2.2.1"
]

[project.optional-dependencies]
dev = [
  "pytest==7.4.3",
  "pytest-asyncio==0.21.1"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"